% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyScript.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers together}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A numeric value}

\item{y}{A numeric value}
}
\value{
The sum of x and y
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
add_numbers(10, -5)
}
