% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_mat_example.R
\docType{data}
\name{dis_mat_example}
\alias{dis_mat_example}
\title{Dissimilarity Matrix Example}
\format{
An object of class \code{\link[stats:dist]{stats::dist()}} containing pairwise similarities for (\eqn{n = 50}) objects.
}
\source{
Generated by simulating \eqn{n = 50} objects from a two-component mixture distribution and then computing a dissimilarity matrix via \code{\link[stats:dist]{stats::dist()}}.
}
\usage{
data(dis_mat_example)
}
\description{
A dissimilarity matrix computed with \code{\link[stats:dist]{stats::dist()}} on a simulated dataset.
}
\details{
The object is intended for examples and vignettes.
}
\keyword{datasets}
