% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeDiagonalSigma}
\alias{makeDiagonalSigma}
\title{Create a diagonal covariance matrix}
\usage{
makeDiagonalSigma(tau_sq_vec)
}
\arguments{
\item{tau_sq_vec}{A vector of length \code{p} containing variance parameters for the covariance matrix.}
}
\value{
A p x p spherical covariance matrix.
}
\description{
A nimbleFunction that returns a diagonal matrix with diagonal entries equal to tau_sq_vec. This function is not intended to be called by users of the package directly.
}
\seealso{
\code{\link[nimble]{nimbleFunction}} for information on nimbleFunctions.
}
\keyword{internal}
