% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{makeSphericalSigma}
\alias{makeSphericalSigma}
\title{Create a spherical covariance matrix}
\usage{
makeSphericalSigma(tau_sq, p)
}
\arguments{
\item{tau_sq}{The variance parameter for the spherical covariance matrix.}

\item{p}{The dimension of the covariance matrix.}
}
\value{
A p x p spherical covariance matrix.
}
\description{
A nimbleFunction that returns a diagonal matrix with all diagonal entries equal to tau_sq. This function is not intended to be called by users of the package directly.
}
\examples{
makeSphericalSigma(tau_sq = 0.1, p = 3)
}
\seealso{
\code{\link[nimble]{nimbleFunction}} for information on nimbleFunctions.
}
\keyword{internal}
