% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mcmc.R
\name{model_mcmc}
\alias{model_mcmc}
\title{MCMC Sampling for a Specified Nimble Model}
\usage{
model_mcmc(
  model_name,
  data,
  constants,
  init_params,
  WAIC,
  nchains,
  niter,
  nburn,
  output_params,
  ...
)
}
\arguments{
\item{data}{A symmetric dissimilarity matrix.}

\item{constants}{A named list of model constant values (including hyperparameters).}

\item{init_params}{A named list of initial values for model parameters.}

\item{nchains}{The number of MCMC chains to run.}

\item{niter}{The number of MCMC iterations to run.}

\item{nburn}{The number of MCMC burn-in iterations.}

\item{...}{Additional arguments passed to \code{nimbleMCMC()}.}

\item{monitors}{A character vector of model parameters to save in the output.}
}
\value{
MCMC samples as a \code{coda} object.
}
\description{
An internal function to run the unequal unrestricted (UU) model. This function is called by \code{\link[=run_dpcd]{run_dpcd()}}.
}
\keyword{internal}
