% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_uu.R
\name{run_uu}
\alias{run_uu}
\title{Run the Unequal Unrestricted Model}
\usage{
run_uu(data, hyper_params, init_params, nchains, niter, nburn, monitors, ...)
}
\arguments{
\item{data}{A symmetric dissimilarity matrix.}

\item{hyper_params}{A named list of hyperparameter values.}

\item{init_params}{A named list of initial values for model parameters.}

\item{nchains}{The number of MCMC chains to run.}

\item{niter}{The number of MCMC iterations to run.}

\item{nburn}{The number of MCMC burn-in iterations.}

\item{monitors}{A character vector of model parameters to save in the output.}

\item{...}{Additional arguments passed to \code{nimbleMCMC()}.}
}
\value{
MCMC samples as a \code{coda} object.
}
\description{
An internal function to run the unequal unrestricted (UU) model. This function is called by \code{\link[=run_dpcd]{run_dpcd()}}.
}
\keyword{internal}
