% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_example.R
\docType{data}
\name{mcmc_example}
\alias{mcmc_example}
\title{MCMC Example Output}
\format{
An object of class \code{mcmc} containing posterior draws for the monitored parameters from the DPCD model fit. It contains 40,000 rows (MCMC iterations) and 125 columns (model parameters).
}
\source{
Generated by fitting an Equal Spherical (ES) DPCD model to the dissimilarities calculated from a simulated dataset (\eqn{n = 50}) with two mixture components.
}
\usage{
data(mcmc_example)
}
\description{
Posterior samples returned by \code{\link[=run_dpcd]{run_dpcd()}} after fitting an Equal Spherical (ES) model to simulated dissimilarities.
}
\details{
The object is intended for examples and vignettes.
}
\keyword{datasets}
