% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive.R
\name{prior_predictive}
\alias{prior_predictive}
\title{Prior Predictive Check}
\usage{
prior_predictive(
  model_name = c("UU", "EU", "UD", "ED", "US", "ES"),
  dis_matrix,
  p,
  trunc_value = 10,
  hyper_params = NULL,
  init_params = NULL,
  scale = TRUE,
  nsim = 1000,
  plot = TRUE
)
}
\arguments{
\item{dis_matrix}{A distance structure such as that returned by \link[stats:dist]{stats::dist} or a full symmetric matrix containing the dissimilarities.}

\item{p}{The dimension of the space in which the objects are embedded. Must be at least 2.}

\item{trunc_value}{The truncation level for the stick-breaking representation of the Dirichlet process.}

\item{hyper_params}{A named list of hyperparameter values. See details for more information.}

\item{init_params}{A named list of initial values for model parameters. See details for more information.}

\item{scale}{Logical indicating whether to scale the distance matrix so that its maximum value is 1.}

\item{nsim}{Number of datasets to simulate from the prior predictive distribution.}

\item{plot}{Logical indicating whether to plot the simulated datasets. See details for more information.}
}
\description{
Prior Predictive Check
}
