% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dpcd.R
\name{run_dpcd}
\alias{run_dpcd}
\title{Run Dirichlet Process Clustering with Dissimilarities}
\usage{
run_dpcd(
  dis_matrix,
  p,
  trunc_value = 10,
  hyper_params,
  init_params,
  output_params,
  scale = TRUE,
  nchains = 1,
  niter = 10000,
  nburn = 0,
  ...
)
}
\arguments{
\item{dis_matrix}{A distance structure such as that returned by \link[stats:dist]{stats::dist} or a full symmetric matrix containing the dissimilarities.}

\item{p}{The dimension of the space in which the objects are embedded. Must be at least 2.}

\item{trunc_value}{The truncation level for the stick-breaking representation of the Dirichlet process.}

\item{hyper_params}{A named list of hyperparameter values. See details for more information.}

\item{init_params}{A named list of initial values for model parameters. See details for more information.}

\item{output_params}{A character vector of model parameters to save in the output. See details for more information.}

\item{scale}{Logical indicating whether to scale the distance matrix so that its maximum value is 1.}

\item{nchains}{Number of MCMC chains to run.}

\item{niter}{Number of MCMC iterations to run.}

\item{nburn}{Number of MCMC burn-in iterations.}

\item{...}{Additional arguments passed to \code{\link[nimble:nimbleMCMC]{nimble::nimbleMCMC()}} from the \code{nimble} package.}
}
\value{
Test
}
\description{
This function fits an infinite mixture model to dissimilarity data using a Dirichlet Process prior. The model is constructed and MCMC sampling is performed using the \href{https://r-nimble.org/}{\code{nimble}} package.
}
\details{
Test
}
\examples{
x <- matrix(rnorm(10*2), ncol = 2)
dis_matrix <- dist(x)
mcmc_samples <- run_dpcd(dis_matrix, p = 2, niter = 1000, nburn = 200)

}
